requirements.txt
# Core Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9

# Data Processing
pandas>=2.2.0
numpy>=1.26.3

# Utilities
python-dotenv>=1.0.0
python-multipart>=0.0.6
requests>=2.31.0
python-dateutil==2.8.2
